modify.client:빈 이미지 주소가 노출되는 문제해결 / modify 강사님코드 처럼 고침
