+++ frontend/src/app/admin/products/modify/[id]/ClientPage.tsx
@@ -1,13 +1,21 @@
 'use client'
 
+import type { components } from '@/lib/backend/apiV1/schema'
 import client from '@/lib/backend/client'
 import { useRouter } from 'next/navigation'
-export default function ClientPage() {
+
+export default function ClientPage({
+  id,
+  responseBody,
+}: {
+  id: string
+  responseBody: components['schemas']['ApiResponseProductDto']
+}) {
   const router = useRouter()
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault()
-    const form = e.target as HTMLFormElement
     let url
+    const form = e.target as HTMLFormElement
     if (form.productName.value.length === 0) {
       alert('이름을 입력해주세요.')
       form.productName.focus()
@@ -24,13 +32,17 @@ export default function ClientPage() {
       return
     }
     if (form.imageUrl.value.length === 0) {
-      url = "https://res.cloudinary.com/heyset/image/upload/v1689582418/buukmenow-folder/no-image-icon-0.jpg"
-
+      url=
+        "https://res.cloudinary.com/heyset/image/upload/v1689582418/buukmenow-folder/no-image-icon-0.jpg"
     }
-    else{
-        url = form.imageUrl.value;
-      }
-    const response = await client.POST('/products', {
+  else
+      url = form.imageUrl.value;
+    const response = await client.PUT('/products/{id}', {
+      params: {
+        path: {
+          id: Number(id),
+        },
+      },
       body: {
         productName: form.productName.value,
         category: form.category.value,
@@ -43,13 +55,12 @@ export default function ClientPage() {
       return
     }
     alert(response.data.message)
-
-    router.replace('/admin/products')
+    router.push('/admin/products/')
   }
   return (
     <>
       <h2 className="text-5xl font-extrabold mt-20 mb-10 text-center">
-        상품 등록
+        상품 수정
       </h2>
       <form onSubmit={handleSubmit}>
         <div className="max-w-[1200px] mx-auto">
@@ -69,6 +80,7 @@ export default function ClientPage() {
                     name="productName"
                     id="productName"
                     className="p-2 h-[50px] w-full border-[1px] border-[#ddd]"
+                    defaultValue={responseBody.content?.productName}
                     placeholder="제품 이름"
                   />
                 </td>
@@ -83,6 +95,7 @@ export default function ClientPage() {
                     name="category"
                     id="category"
                     className="p-2 h-[50px] w-full border-[1px] border-[#ddd]"
+                    defaultValue={responseBody.content?.category}
                     placeholder="카테고리"
                   />
                 </td>
@@ -97,6 +110,7 @@ export default function ClientPage() {
                     name="price"
                     id="price"
                     className="p-2 h-[50px] w-full border-[1px] border-[#ddd]"
+                    defaultValue={responseBody.content?.price}
                     placeholder="가격"
                   />
                 </td>
@@ -111,6 +125,7 @@ export default function ClientPage() {
                     name="imageUrl"
                     id="imageUrl"
                     className="p-2 h-[50px] w-full border-[1px] border-[#ddd]"
+                    defaultValue={responseBody.content?.imageUrl}
                     placeholder="이미지"
                   />
                 </td>
